// src/types/firestore.ts
export interface FirestoreCategory {
  id: string;
  name: string;
  slug: string;
  image: string; // Cloudinary URL
  order: number;
  isActive: boolean;
  createdAt: Timestamp | null;
}